Tests/LibWeb/Text/input/credential-protection-trusted-form.html

<!DOCTYPE html>
<html>
<head>
    <title>Credential Protection: Trusted Form Workflow Test</title>
    <script src="include.js"></script>
</head>
<body>
    <h1>Trusted Form Workflow Test</h1>

    <!-- Cross-origin form that should become trusted after user approval -->
    <form id="trustedForm" action="https://accounts.example.com/authenticate" method="POST">
        <input type="text" name="email" id="email" value="user@example.com" />
        <input type="password" name="password" id="password" value="mypassword" />
        <button type="submit" id="submitBtn">Sign In</button>
    </form>

    <script>
        test(() => {
            println("=== Credential Protection: Trusted Form Workflow Test ===");
            println();

            let form = document.getElementById('trustedForm');
            let formActionURL = new URL(form.action);
            let currentOrigin = window.location.origin;

            // Test 1: Initial form state
            println("Test 1: Form configuration");
            println("  form exists: " + (form !== null));
            println("  form action: " + form.action);
            println("  form method: " + form.method.toUpperCase());
            println("  current origin: " + currentOrigin);
            println("  action origin: " + formActionURL.origin);
            println();

            // Test 2: Cross-origin characteristics
            println("Test 2: Cross-origin analysis");
            let currentHost = window.location.hostname || 'localhost';
            let actionHost = formActionURL.hostname;
            println("  current host: " + currentHost);
            println("  action host: " + actionHost);
            println("  is cross-origin: " + (actionHost !== currentHost && actionHost !== ''));
            println("  uses HTTPS: " + (formActionURL.protocol === 'https:'));
            println();

            // Test 3: Credential fields
            let passwordField = document.getElementById('password');
            let emailField = document.getElementById('email');
            println("Test 3: Credential fields");
            println("  has password field: " + (passwordField !== null));
            println("  password type: " + passwordField.type);
            println("  has email field: " + (emailField !== null));
            println("  email field name: " + emailField.name);
            println();

            // Test 4: Trusted form workflow stages
            println("Test 4: Trusted form workflow stages");
            println();
            println("  Stage 1: INITIAL SUBMISSION");
            println("    - FormMonitor detects cross-origin submission with password");
            println("    - Alert is triggered: credential_exfiltration or third_party_form_post");
            println("    - SecurityAlertDialog shown to user");
            println("    - User has options: Block, Trust, Learn More");
            println();
            println("  Stage 2: USER TRUSTS THE FORM");
            println("    - User clicks 'Trust' button in SecurityAlertDialog");
            println("    - FormMonitor.grant_trusted_relationship() is called");
            println("    - Trusted relationship stored in m_trusted_relationships HashMap");
            println("    - Key: form_origin -> action_origin");
            println("    - Form submission proceeds");
            println();
            println("  Stage 3: SUBSEQUENT SUBMISSIONS");
            println("    - FormMonitor checks m_trusted_relationships before alerting");
            println("    - Finds existing trust relationship");
            println("    - NO alert is triggered");
            println("    - Form submission proceeds silently");
            println();
            println("  Stage 4: TRUST MANAGEMENT");
            println("    - User can view trusted relationships in about:security");
            println("    - User can revoke trust via 'Revoke Trust' button");
            println("    - Revocation removes entry from m_trusted_relationships");
            println("    - Future submissions will trigger alerts again");
            println();

            // Test 5: FormMonitor data structure
            println("Test 5: FormMonitor internal state");
            println("  Expected data structures:");
            println("  - m_trusted_relationships: HashMap<ByteString, Vector<ByteString>>");
            println("  - Key format: form_origin (serialized)");
            println("  - Value format: Vector of trusted action_origins");
            println("  - Example: 'about:blank' -> ['https://accounts.example.com']");
            println();

            // Test 6: Security considerations
            println("Test 6: Security considerations");
            println("  - Trust is origin-specific, not URL-specific");
            println("  - HTTPS requirement enforced for trusted submissions");
            println("  - Trust persists only in memory (not saved to disk)");
            println("  - Trust cleared when browser restarts");
            println("  - User maintains full control via about:security");
            println();

            // Test 7: IPC message flow
            println("Test 7: IPC message flow for trust workflow");
            println("  1. WebContent -> UI: on_credential_exfiltration_alert()");
            println("  2. UI shows SecurityAlertDialog");
            println("  3. User clicks 'Trust'");
            println("  4. UI -> WebContent: async_credential_alert_action('trust')");
            println("  5. WebContent calls FormMonitor.grant_trusted_relationship()");
            println("  6. FormMonitor stores trust in m_trusted_relationships");
            println("  7. Form submission proceeds");
            println();

            println("=== Expected Test Outcomes ===");
            println("First submission:");
            println("  ✓ Alert triggered");
            println("  ✓ User prompted for decision");
            println("  ✓ Trust can be granted");
            println();
            println("After trusting:");
            println("  ✓ Subsequent submissions silent");
            println("  ✓ No alerts for same form-action pair");
            println("  ✓ Trust visible in about:security");
            println();
            println("After revoking trust:");
            println("  ✓ Alerts resume for this form");
            println("  ✓ User must re-trust if desired");
            println();

            println("=== Test Complete ===");
            println("This test documents the trusted form relationship workflow.");
        });
    </script>
</body>
</html>
