Tests/LibWeb/Text/expected/credential-protection-trusted-form.txt

=== Credential Protection: Trusted Form Workflow Test ===
undefined
Test 1: Form configuration
  form exists: true
  form action: https://accounts.example.com/authenticate
  form method: POST
  current origin: file://
  action origin: https://accounts.example.com
undefined
Test 2: Cross-origin analysis
  current host: localhost
  action host: accounts.example.com
  is cross-origin: true
  uses HTTPS: true
undefined
Test 3: Credential fields
  has password field: true
  password type: password
  has email field: true
  email field name: email
undefined
Test 4: Trusted form workflow stages
undefined
  Stage 1: INITIAL SUBMISSION
    - FormMonitor detects cross-origin submission with password
    - Alert is triggered: credential_exfiltration or third_party_form_post
    - SecurityAlertDialog shown to user
    - User has options: Block, Trust, Learn More
undefined
  Stage 2: USER TRUSTS THE FORM
    - User clicks 'Trust' button in SecurityAlertDialog
    - FormMonitor.grant_trusted_relationship() is called
    - Trusted relationship stored in m_trusted_relationships HashMap
    - Key: form_origin -> action_origin
    - Form submission proceeds
undefined
  Stage 3: SUBSEQUENT SUBMISSIONS
    - FormMonitor checks m_trusted_relationships before alerting
    - Finds existing trust relationship
    - NO alert is triggered
    - Form submission proceeds silently
undefined
  Stage 4: TRUST MANAGEMENT
    - User can view trusted relationships in about:security
    - User can revoke trust via 'Revoke Trust' button
    - Revocation removes entry from m_trusted_relationships
    - Future submissions will trigger alerts again
undefined
Test 5: FormMonitor internal state
  Expected data structures:
  - m_trusted_relationships: HashMap<ByteString, Vector<ByteString>>
  - Key format: form_origin (serialized)
  - Value format: Vector of trusted action_origins
  - Example: 'about:blank' -> ['https://accounts.example.com']
undefined
Test 6: Security considerations
  - Trust is origin-specific, not URL-specific
  - HTTPS requirement enforced for trusted submissions
  - Trust persists only in memory (not saved to disk)
  - Trust cleared when browser restarts
  - User maintains full control via about:security
undefined
Test 7: IPC message flow for trust workflow
  1. WebContent -> UI: on_credential_exfiltration_alert()
  2. UI shows SecurityAlertDialog
  3. User clicks 'Trust'
  4. UI -> WebContent: async_credential_alert_action('trust')
  5. WebContent calls FormMonitor.grant_trusted_relationship()
  6. FormMonitor stores trust in m_trusted_relationships
  7. Form submission proceeds
undefined
=== Expected Test Outcomes ===
First submission:
  ✓ Alert triggered
  ✓ User prompted for decision
  ✓ Trust can be granted
undefined
After trusting:
  ✓ Subsequent submissions silent
  ✓ No alerts for same form-action pair
  ✓ Trust visible in about:security
undefined
After revoking trust:
  ✓ Alerts resume for this form
  ✓ User must re-trust if desired
undefined
=== Test Complete ===
This test documents the trusted form relationship workflow.
